Detection/2 caméras/controller/requirements.txt
ackermann-msgs==2.0.2
action-msgs==1.2.1
action-tutorials-interfaces==0.20.5
action-tutorials-py==0.20.5
actionlib-msgs==4.2.4
actuator-msgs==0.0.1
ament-cmake-google-benchmark==1.3.10
ament-cmake-test==1.3.10
ament-copyright==0.12.11
ament-cppcheck==0.12.11
ament-cpplint==0.12.11
ament-flake8==0.12.11
ament-index-python==1.4.0
ament-lint==0.12.11
ament-lint-cmake==0.12.11
ament-package==0.14.0
ament-pep257==0.12.11
ament-uncrustify==0.12.11
ament-xmllint==0.12.11
angles==1.15.0
backports.ssl-match-hostname==3.7.0.1
bond==3.0.2
builtin-interfaces==1.2.1
camera-calibration==3.0.6
cartographer-ros-msgs==2.0.9002
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==3.4.1
composition-interfaces==1.2.1
contourpy==1.3.1
control-msgs==4.6.0
controller-manager==2.43.1
controller-manager-msgs==2.43.1
cv-bridge==3.2.1
cycler==0.12.1
demo-nodes-py==0.20.5
diagnostic-msgs==4.2.4
diagnostic-updater==4.0.0
domain-coordinator==0.10.0
dwb-msgs==1.1.16
example-interfaces==0.9.3
examples-rclpy-executors==0.15.2
examples-rclpy-minimal-action-client==0.15.2
examples-rclpy-minimal-action-server==0.15.2
examples-rclpy-minimal-client==0.15.2
examples-rclpy-minimal-publisher==0.15.2
examples-rclpy-minimal-service==0.15.2
examples-rclpy-minimal-subscriber==0.15.2
filelock==3.16.1
fonttools==4.55.3
fsspec==2024.12.0
gazebo-msgs==3.7.0
generate-parameter-library-py==0.3.8
geometry-msgs==4.2.4
gps-msgs==2.0.4
graph-msgs==0.2.0
idna==2.10
image-geometry==3.2.1
interactive-markers==2.3.2
Jinja2==3.1.5
joint-state-publisher==2.4.0
joint-state-publisher-gui==2.4.0
kiwisolver==1.4.8
laser-geometry==2.4.0
launch==1.0.6
launch-param-builder==0.1.1
launch-ros==0.19.7
launch-testing==1.0.6
launch-testing-ros==0.19.7
launch-xml==1.0.6
launch-yaml==1.0.6
lifecycle-msgs==1.2.1
logging-demo==0.20.5
map-msgs==2.1.0
MarkupSafe==3.0.2
matplotlib==3.10.0
message-filters==4.3.5
moveit-configs-utils==2.5.5
moveit-msgs==2.2.1
mpmath==1.3.0
nav-2d-msgs==1.1.16
nav-msgs==4.2.4
nav2-common==1.1.16
nav2-msgs==1.1.16
nav2-simple-commander==1.0.0
networkx==3.4.2
numpy==2.2.1
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
object-recognition-msgs==2.0.0
octomap-msgs==2.0.0
opencv-python==4.10.0.84
osrf-pycommon==2.0.2
packaging==24.2
paho-mqtt==2.1.0
pandas==2.2.3
pcl-msgs==1.0.0
pendulum-msgs==0.20.5
pillow==11.1.0
psutil==6.1.1
py-cpuinfo==9.0.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-qt-binding==1.1.2
pytz==2024.2
PyYAML==6.0.2
qt-dotgraph==2.2.3
qt-gui==2.2.3
qt-gui-cpp==2.2.3
qt-gui-py-common==2.2.3
quality-of-service-demo-py==0.20.5
rcl-interfaces==1.2.1
rclpy==3.3.14
rcutils==5.1.6
requests==2.27.1
resource-retriever==3.1.2
rmw-dds-common==1.6.0
ros-gz-interfaces==0.244.15
ros-ign-interfaces==0.244.15
ros2action==0.18.11
ros2bag==0.15.12
ros2cli==0.18.11
ros2component==0.18.11
ros2controlcli==2.43.1
ros2doctor==0.18.11
ros2interface==0.18.11
ros2launch==0.19.7
ros2lifecycle==0.18.11
ros2multicast==0.18.11
ros2node==0.18.11
ros2param==0.18.11
ros2pkg==0.18.11
ros2run==0.18.11
ros2service==0.18.11
ros2test==0.4.0
ros2topic==0.18.11
rosbag2-interfaces==0.15.12
rosbag2-py==0.15.12
rosgraph-msgs==1.2.1
rosidl-adapter==3.1.5
rosidl-cli==3.1.5
rosidl-cmake==3.1.5
rosidl-generator-c==3.1.5
rosidl-generator-cpp==3.1.5
rosidl-generator-py==0.14.4
rosidl-parser==3.1.5
rosidl-runtime-py==0.9.3
rosidl-typesupport-c==2.0.1
rosidl-typesupport-cpp==2.0.1
rosidl-typesupport-fastrtps-c==2.2.2
rosidl-typesupport-fastrtps-cpp==2.2.2
rosidl-typesupport-introspection-c==3.1.5
rosidl-typesupport-introspection-cpp==3.1.5
rpyutils==0.2.1
rqt-action==2.0.1
rqt-bag==1.1.5
rqt-bag-plugins==1.1.5
rqt-console==2.0.3
rqt-graph==1.3.1
rqt-gui==1.1.7
rqt-gui-py==1.1.7
rqt-msg==1.2.0
rqt-plot==1.1.2
rqt-publisher==1.5.0
rqt-py-common==1.1.7
rqt-py-console==1.0.2
rqt-reconfigure==1.1.2
rqt-service-caller==1.0.5
rqt-shell==1.0.2
rqt-srv==1.0.3
rqt-topic==1.5.0
scipy==1.15.0
scripts==3.7.0
seaborn==0.13.2
sensor-msgs==4.2.4
sensor-msgs-py==4.2.4
shape-msgs==4.2.4
simplejson==3.16.0
six==1.17.0
slam-toolbox==2.6.8
smclib==3.0.2
srdfdom==2.0.5
sros2==0.10.5
statistics-msgs==1.2.1
std-msgs==4.2.4
std-srvs==4.2.4
stereo-msgs==4.2.4
sympy==1.13.1
teleop-twist-keyboard==2.4.0
tf-transformations==1.1.0
tf2-geometry-msgs==0.25.8
tf2-kdl==0.25.8
tf2-msgs==0.25.8
tf2-py==0.25.8
tf2-ros-py==0.25.8
tf2-tools==0.25.8
theora-image-transport==2.5.2
topic-monitor==0.20.5
torch==2.5.1
torchvision==0.20.1
TornadIO2==0.0.4
tornado==3.2.2
tqdm==4.67.1
trajectory-msgs==4.2.4
triton==3.1.0
turtle-tf2-py==0.3.7
turtlebot3-example==2.1.5
turtlebot3-msgs==2.2.3
turtlebot3-teleop==2.1.5
turtlesim==1.4.2
typing_extensions==4.12.2
tzdata==2024.2
ultralytics==8.3.58
ultralytics-thop==2.0.13
unique-identifier-msgs==2.2.1
urdfdom-py==1.2.1
urllib3==1.26.18
vision-msgs==4.1.1
visualization-msgs==4.2.4
xacro==2.0.8
